Fix when ifcOWL is missing GUID-values
In the case when an ifcOWL file is missing GUID triples and is still used for the conversion.  The fix corrects the corrupted input file by generating new GUIDs and then converts the triple set into LBD RDF.

As there may be more errors contained  in the input, corrupted input files shouldn't actually be allowed.